% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Proxy.R
\name{close_db_connection}
\alias{close_db_connection}
\title{Close a Database Connection}
\usage{
close_db_connection(conn)
}
\arguments{
\item{conn}{A database connection object to be closed.}
}
\description{
This function closes the specified database connection.
}
\examples{
close_db_connection(conn)
}
