% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_player_display.R
\name{create_percentiles_plot}
\alias{create_percentiles_plot}
\title{Create a Custom Plotly Visualization}
\usage{
create_percentiles_plot(plot_data, addition)
}
\arguments{
\item{plot_data}{A data frame containing the data to be visualized. It should include `metric`, `percentile`, and `value` columns.}

\item{addition}{A string to append to the y-axis title and hover text for additional context.}
}
\value{
A Plotly object representing the custom visualization.
}
\description{
Generates a Plotly visualization with segments and markers to display metrics and percentiles.
}
