% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{create_table}
\alias{create_table}
\title{Create a Table in the Database (Structure Only)}
\usage{
create_table(conn, table_name, data, index_cols = NULL, override = FALSE)
}
\arguments{
\item{conn}{An active database connection object.}

\item{table_name}{A string representing the name of the table to create.}

\item{data}{A data frame (or tibble) containing the data structure to create the table.}

\item{index_cols}{A character vector of column names to create an index on (optional).}

\item{override}{A logical value (TRUE/FALSE). If TRUE, the existing table is dropped and replaced with the new one.}
}
\description{
This function creates a table in the database based on the structure of the provided data (without inserting any data).
If the table already exists, it will not stop execution.
}
