% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Proxy.R
\name{delete_db_table}
\alias{delete_db_table}
\title{Delete a Database Table}
\usage{
delete_db_table(table_name)
}
\arguments{
\item{table_name}{A string representing the name of the table to be deleted.}

\item{conn}{A database connection object.}
}
\value{
A character string indicating whether the operation was successful.
}
\description{
This function deletes the specified table from the database.
}
\examples{
delete_db_table(conn, "games")
}
