% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_Game_Processing.R
\name{Endpoint}
\alias{Endpoint}
\title{Endpoint Class}
\description{
The `Endpoint` class enables you to interact with an API by constructing a complete URL based on a base URL and endpoint. 
It includes methods for fetching data from the URL and processing it into data frames.
}
\details{
This class is used to manage the fetching of data from a specified API endpoint.
It allows for setting a base URL, constructing a complete URL, and fetching data in the form of data frames.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{A character string representing the base URL for the API}

\item{\code{endpoint}}{A character string representing the specific endpoint}

\item{\code{url}}{A character string representing the full URL to the API resource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Endpoint-new}{\code{Endpoint$new()}}
\item \href{#method-Endpoint-fetch_dfs_from_url}{\code{Endpoint$fetch_dfs_from_url()}}
\item \href{#method-Endpoint-get_url}{\code{Endpoint$get_url()}}
\item \href{#method-Endpoint-get_endpoint}{\code{Endpoint$get_endpoint()}}
\item \href{#method-Endpoint-clone}{\code{Endpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Endpoint object with a given base URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$new(base_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{A character string representing the base URL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-fetch_dfs_from_url"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-fetch_dfs_from_url}{}}}
\subsection{Method \code{fetch_dfs_from_url()}}{
Fetches data from the constructed URL and returns it as a list of data frames.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$fetch_dfs_from_url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of data frames fetched from the URL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-get_url}{}}}
\subsection{Method \code{get_url()}}{
Constructs and returns the full URL by combining the base URL and endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$get_url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the full URL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-get_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-get_endpoint}{}}}
\subsection{Method \code{get_endpoint()}}{
This is a placeholder method for setting the endpoint.
In a real implementation, this would be used to specify the API endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$get_endpoint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
