% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_proxy.R
\name{fetch_games}
\alias{fetch_games}
\title{Fetch Game Data from the API and Store in Database}
\usage{
fetch_games(base_url, endpoint = "games?date=2021:")
}
\value{
A character string indicating whether the operation was successful or failed.
}
\description{
This function fetches game data from a specified API endpoint and stores it in a local SQLite database.
It makes an HTTP request to the API, parses the data, and writes the relevant game data to the database table.
}
