% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_proxy.R
\name{fetch_player_stats}
\alias{fetch_player_stats}
\title{Fetch Player Stats Data from the API and Store in Database}
\usage{
fetch_player_stats(base_url, players)
}
\value{
A character string indicating whether the operation was successful or failed.
}
\description{
This function fetches player data from a specified API endpoint and stores it in a local SQLite database.
It makes an HTTP request to the API, parses the data, and writes the relevant player data to the database table.
}
