% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_proxy.R
\name{fetch_teams}
\alias{fetch_teams}
\title{Fetch Game Data from the API and Store in Database}
\usage{
fetch_teams(base_url, years = NULL, teamID = NULL)
}
\value{
A character string indicating whether the operation was successful or failed.
}
\description{
This function fetches player data from a specified API endpoint
}
