% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_Proxy.R
\name{GameEventsProxy}
\alias{GameEventsProxy}
\title{GameEventsProxy Class}
\description{
The `GameEventsProxy` class is a subclass of `Endpoint`, designed to interact with the API endpoint for game events.
It includes methods for fetching throws, pulls, and penalties from a game based on its `gameID`.
}
\details{
This class extends the `Endpoint` class to provide specific functionality for
fetching and processing game event data from a given API.
}
\section{Super class}{
\code{\link[ultistats:Endpoint]{ultistats::Endpoint}} -> \code{GameEventsProxy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GameEventsProxy-new}{\code{GameEventsProxy$new()}}
\item \href{#method-GameEventsProxy-get_request}{\code{GameEventsProxy$get_request()}}
\item \href{#method-GameEventsProxy-get_throws_from_id}{\code{GameEventsProxy$get_throws_from_id()}}
\item \href{#method-GameEventsProxy-get_pulls_from_id}{\code{GameEventsProxy$get_pulls_from_id()}}
\item \href{#method-GameEventsProxy-get_penalties_from_id}{\code{GameEventsProxy$get_penalties_from_id()}}
\item \href{#method-GameEventsProxy-clone}{\code{GameEventsProxy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ultistats" data-topic="Endpoint" data-id="fetch_dfs_from_url"><a href='../../ultistats/html/Endpoint.html#method-Endpoint-fetch_dfs_from_url'><code>ultistats::Endpoint$fetch_dfs_from_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ultistats" data-topic="Endpoint" data-id="get_endpoint"><a href='../../ultistats/html/Endpoint.html#method-Endpoint-get_endpoint'><code>ultistats::Endpoint$get_endpoint()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ultistats" data-topic="Endpoint" data-id="get_url"><a href='../../ultistats/html/Endpoint.html#method-Endpoint-get_url'><code>ultistats::Endpoint$get_url()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-new}{}}}
\subsection{Method \code{new()}}{
Initializes the GameEventsProxy object, setting the base URL for game event API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-get_request"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-get_request}{}}}
\subsection{Method \code{get_request()}}{
Makes a GET request to the API and stores the response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$get_request(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The specific endpoint to query.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-get_throws_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-get_throws_from_id}{}}}
\subsection{Method \code{get_throws_from_id()}}{
Fetches the throws from a game by gameID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$get_throws_from_id(gameID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameID}}{The ID of the game for which to fetch the throws.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble containing the throw events for the game.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-get_pulls_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-get_pulls_from_id}{}}}
\subsection{Method \code{get_pulls_from_id()}}{
Fetches the pulls from a game by gameID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$get_pulls_from_id(gameID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameID}}{The ID of the game for which to fetch the pulls.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble containing the pull events for the game.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-get_penalties_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-get_penalties_from_id}{}}}
\subsection{Method \code{get_penalties_from_id()}}{
Fetches the penalties from a game by gameID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$get_penalties_from_id(gameID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameID}}{The ID of the game for which to fetch the penalties.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector containing the count of home and away team penalties.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameEventsProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GameEventsProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEventsProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
