% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_player_display.R
\name{get_filtered_throws}
\alias{get_filtered_throws}
\title{Retrieve and Process Player Throws}
\usage{
get_filtered_throws(db_path, player_id)
}
\arguments{
\item{db_path}{Path to the database file.}

\item{player_id}{Player's ID.}
}
\value{
A data frame with processed throw data, including `adjusted_angle` and `year`.
}
\description{
Fetches throw data for a player from the database and processes it to add adjusted angles and year.
}
\examples{
\dontrun{
  get_filtered_throws("path/to/database.sqlite", 12345)
}
}
