% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_Update_Throws_Table.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Game IDs Based on Date Range}
\usage{
get_game_ids(start_date = "2021-01-01", end_date = Sys.Date())
}
\arguments{
\item{start_date}{A string representing the start date of the date range (default is "2021-01-01").
The date should be in the format "YYYY-MM-DD".}

\item{end_date}{A string representing the end date of the date range (default is the current date).
The date should be in the format "YYYY-MM-DD".}
}
\value{
A character vector containing the game IDs.
        If the request is unsuccessful, an error is raised.
}
\description{
This function queries the API to retrieve a list of game IDs based on a specified date range.
It constructs the URL for the API request using the provided start and end dates, sends the request,
and returns the game IDs from the response.
}
