% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{get_recent_timestamp}
\alias{get_recent_timestamp}
\title{Get the Most Recent Timestamp from a Table}
\usage{
get_recent_timestamp(conn, table)
}
\arguments{
\item{conn}{An active database connection object.}

\item{table}{A string representing the name of the table.}
}
\value{
The most recent timestamp (as POSIXct object).
}
\description{
This function retrieves the most recent timestamp (from the insertTimestamp column) 
in a specific table.
}
