% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{get_table}
\alias{get_table}
\title{Get Throws Data}
\usage{
get_table(conn, table_name)
}
\arguments{
\item{conn}{An active database connection object.}

\item{table_name}{A string representing the name of the table (default is 'throws').}
}
\value{
A data frame containing all the data from the throws table.
}
\description{
This function retrieves all the data from the throws table in the database.
}
