% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{get_table_names}
\alias{get_table_names}
\title{Get Table Names from the Database}
\usage{
get_table_names(conn)
}
\arguments{
\item{conn}{An active database connection object.}
}
\value{
A vector of table names (character vector).
}
\description{
This function retrieves the names of all tables present in the SQLite database.
}
