% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_admin_data.R
\name{interpolate_time_left}
\alias{interpolate_time_left}
\title{Interpolate Time Left}
\usage{
interpolate_time_left(data, next_start_time)
}
\description{
Interpolate time left based on the number of rows in a group. 
creates linear interpolation based on time left to start the point, time left 
starting the next point, and the number of throws in the point
}
