% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Proxy.R
\name{open_db_connection}
\alias{open_db_connection}
\title{Open a Database Connection}
\usage{
open_db_connection(db_path = NULL)
}
\arguments{
\item{db_path}{A string representing the path to the SQLite database.}
}
\value{
A database connection object.
}
\description{
This function opens a connection to an SQLite database specified by the given path.
}
\examples{
conn <- open_db_connection("path/to/database.db")
}
