% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_player_display.R
\name{radial_histogram_plot}
\alias{radial_histogram_plot}
\title{Create Radial Histogram Plot}
\usage{
radial_histogram_plot(throws, bin_width = 24)
}
\arguments{
\item{throws}{A numeric vector of adjusted throw angles.}

\item{bin_width}{Width of the bins for the histogram in degrees. Default is 24.}
}
\value{
A `ggplot` object representing the radial histogram.
}
\description{
Generates a radial histogram plot for throw angles.
}
