% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{reload_button_ui}
\alias{reload_button_ui}
\title{Create a Reload Button UI Component}
\usage{
reload_button_ui(id, label = "Reload API", icon = NULL)
}
\arguments{
\item{ns}{A function used to namespace the input IDs (required in a modular app).}
}
\value{
A `shiny.tag` object representing the action button UI component.
}
\description{
This function generates a UI component that displays a button to reload an API.
The button is styled with the Bootstrap `btn-outline-primary` class and includes
a sync icon from Font Awesome.
}
