% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_Game_Processing.R
\name{TeamEvents}
\alias{TeamEvents}
\title{TeamEvents Class}
\description{
TeamEvents Class

TeamEvents Class
}
\details{
This class handles the events of a game, such as throws and pulls, and processes them for both home and away teams. It tracks the scores, penalties, and various game events.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TeamEvents-new}{\code{TeamEvents$new()}}
\item \href{#method-TeamEvents-add_event}{\code{TeamEvents$add_event()}}
\item \href{#method-TeamEvents-set_pull_event}{\code{TeamEvents$set_pull_event()}}
\item \href{#method-TeamEvents-add_pull}{\code{TeamEvents$add_pull()}}
\item \href{#method-TeamEvents-end_point}{\code{TeamEvents$end_point()}}
\item \href{#method-TeamEvents-end_quarter}{\code{TeamEvents$end_quarter()}}
\item \href{#method-TeamEvents-clone}{\code{TeamEvents$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-new"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-new}{}}}
\subsection{Method \code{new()}}{
Constructor for TeamEvents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$new(game_events, home_team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{game_events}}{A list of game events to process.}

\item{\code{home_team}}{A logical value indicating whether the team is the home team.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-add_event"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-add_event}{}}}
\subsection{Method \code{add_event()}}{
Add a throw event to the processed events.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$add_event(throw_event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{throw_event}}{A `ThrowEvent` object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-set_pull_event"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-set_pull_event}{}}}
\subsection{Method \code{set_pull_event()}}{
Add a throw event to the processed events.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$set_pull_event(throw_event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{throw_event}}{A `ThrowEvent` object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-add_pull"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-add_pull}{}}}
\subsection{Method \code{add_pull()}}{
Add a pull event to the processed pulls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$add_pull(pull_event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pull_event}}{A `PullEvent` object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-end_point"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-end_point}{}}}
\subsection{Method \code{end_point()}}{
Reset the state at the end of a point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$end_point()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-end_quarter"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-end_quarter}{}}}
\subsection{Method \code{end_quarter()}}{
Reset the state at the end of a quarter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$end_quarter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamEvents-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TeamEvents-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamEvents$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
