% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_Game_Processing.R
\name{ThrowEvent}
\alias{ThrowEvent}
\title{GameEvents Class}
\description{
GameEvents Class

GameEvents Class
}
\details{
This class handles the game events for a specific game by retrieving game information and events,
processing the events, and providing a data frame of the processed events.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThrowEvent-new}{\code{ThrowEvent$new()}}
\item \href{#method-ThrowEvent-clone}{\code{ThrowEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThrowEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThrowEvent-new}{}}}
\subsection{Method \code{new()}}{
Constructor for ThrowEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThrowEvent$new(
  thrower,
  thrower_x,
  thrower_y,
  receiver,
  receiver_x,
  receiver_y,
  turnover
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thrower}}{The player who threw the disc.}

\item{\code{thrower_x}}{The x-coordinate of the thrower.}

\item{\code{thrower_y}}{The y-coordinate of the thrower.}

\item{\code{receiver}}{The player who received the disc.}

\item{\code{receiver_x}}{The x-coordinate of the receiver.}

\item{\code{receiver_y}}{The y-coordinate of the receiver.}

\item{\code{turnover}}{Flag indicating whether the throw resulted in a turnover.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThrowEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThrowEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThrowEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThrowEvent-new}{\code{ThrowEvent$new()}}
\item \href{#method-ThrowEvent-clone}{\code{ThrowEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThrowEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThrowEvent-new}{}}}
\subsection{Method \code{new()}}{
Constructor for ThrowEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThrowEvent$new(
  thrower,
  thrower_x,
  thrower_y,
  receiver,
  receiver_x,
  receiver_y,
  turnover
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thrower}}{The player who threw the disc.}

\item{\code{thrower_x}}{The x-coordinate of the thrower.}

\item{\code{thrower_y}}{The y-coordinate of the thrower.}

\item{\code{receiver}}{The player who received the disc.}

\item{\code{receiver_x}}{The x-coordinate of the receiver.}

\item{\code{receiver_y}}{The y-coordinate of the receiver.}

\item{\code{turnover}}{Flag indicating whether the throw resulted in a turnover.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThrowEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThrowEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThrowEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
