% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_admin_data.R
\name{update_advanced_stats}
\alias{update_advanced_stats}
\title{Function to Process and Update Advanced Stats}
\usage{
update_advanced_stats(pool, base_url)
}
\arguments{
\item{pool}{A database connection object to interact with the database.}

\item{base_url}{The base URL for fetching external data (currently unused in the function).}
}
\value{
NULL This function does not return any value. It updates the `advanced_stats` table in the database.
}
\description{
This function fetches throws data, processes it through prediction models for both feature value (fv) and control points (cp), then applies additional transformations and calculations to derive advanced statistics. Finally, it updates or creates the `advanced_stats` table in the database with the processed data.
}
