% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_admin_data.R
\name{update_player_stats}
\alias{update_player_stats}
\title{Function to Process and Update Player Stats}
\usage{
update_player_stats(pool, base_url)
}
\arguments{
\item{pool}{A database connection object to interact with the database.}

\item{base_url}{The base URL to fetch player data from an external API or service.}
}
\value{
NULL This function does not return any value. It updates the `player_stats` table in the database.
}
\description{
This function fetches, processes, and updates the player statistics in a database. It computes advanced statistics, combines them with yearly and career data, adds timestamps, calculates additional stats, and processes data for final insertion or update in the `player_stats` table.
}
