% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Proxy.R
\name{update_table}
\alias{update_table}
\title{Update Table in Database}
\usage{
update_table(table_name, game_data, game_id_column = "gameID", current_game_id)
}
\arguments{
\item{table_name}{The name of the table to update (e.g., "throws").}

\item{game_data}{The new data to append to the table (data frame).}

\item{game_id_column}{The column name for `gameID` (defaults to "gameID").}

\item{current_game_id}{The value of the `gameID` to delete from the table before appending new data.}

\item{conn}{The database connection object.}
}
\value{
A message indicating whether the operation was successful or not.
}
\description{
A function to update a specified table in the database by removing existing rows with a matching `gameID` and appending new data.
}
