% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Proxy.R
\name{write_db_table}
\alias{write_db_table}
\title{Write Data to a Database Table}
\usage{
write_db_table(table_name, data, overwrite = TRUE)
}
\arguments{
\item{table_name}{A string representing the name of the table.}

\item{data}{A data frame to be written to the table.}

\item{overwrite}{A logical value indicating whether to overwrite the table if it exists (default is TRUE).}

\item{conn}{A database connection object.}
}
\value{
A character string indicating whether the operation was successful.
}
\description{
This function writes data to a specified table in the database.
}
\examples{
write_db_table("games", games_df)
}
