% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{close_db_connection}
\alias{close_db_connection}
\title{Close the Database Connection}
\usage{
close_db_connection(conn)
}
\description{
This function closes the database connection if it is open. If the connection is
not open, it silently returns without doing anything.
}
