% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Unique Game IDs from the Games Table}
\usage{
get_game_ids(conn)
}
\arguments{
\item{conn}{An active database connection object.}
}
\value{
A vector of unique game IDs (character vector).
}
\description{
This function retrieves a list of unique game IDs from the 'games' table.
}
