% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_proxy.R
\name{open_db_connection}
\alias{open_db_connection}
\title{Open a Database Connection}
\usage{
open_db_connection(db_path)
}
\arguments{
\item{db_path}{A string representing the path to the SQLite database file. This is
required for SQLite databases.}
}
\value{
An active database connection object.
}
\description{
This function opens a connection to a database. If the connection is already open,
it does nothing and returns the existing connection.
}
